v1/requirements.txt
# Core LangChain packages
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1

# LLM providers
langchain-google-genai==1.0.3
langchain-openai==0.1.7
langchain-anthropic==0.1.4

# Vector store
langchain-qdrant==0.1.0
qdrant-client==1.7.3

# Document processing
pypdf==4.1.0
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.6
beautifulsoup4==4.12.3
lxml==5.1.0

# Data processing
pandas==2.2.1
numpy==1.26.4
scipy==1.12.0

# Web framework (for API)
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4

# Environment and configuration
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Utilities
tqdm==4.66.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.9

# Logging and monitoring
structlog==24.1.0

# Authentication (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing (optional)
pytest==8.1.1
pytest-asyncio==0.23.5

# Development tools
black==24.3.0
isort==5.13.2
flake8==7.0.0